requirements.txt
pyserial==3.5
opencv-python==4.5.1